Added functionality to the installer's "Choose Terraria Path" button
This button allows the user to choose the Terraria executable for the
installer to back-up and replace with tModLoader. Because OpenFileDialog
does not work in Mac/Linux, there is also the option to drag and drop
the file onto the installer window. The installer will also use a
default file location for the Terraria executable if the user has not
specified one.