New/overloaded methods for wing related methods
Vertical/HorizontalWingSpeeds now pass the player as a parameter (overloaded version calls the old version by default so it doesn't break anything)
ItemLoader.WingUpdate returns a bool value. For other classes, there's a new NewWingUpdate method that returns a bool, and it calls the current WingUpdate to not break anything. Return value of false keeps everything the same as it is, using stock animation logic. Return value of true requires the animation to be handled manually.