Added Better MP3 Support
MP3s are no longer converted to wavs and cached, now they only hold the
part of the mp3 that needs to be played but only for sounds. Normal none
music mp3s are still converted when needing to be played but no caching
is done for any of them reguardless.