Volcano Tremor Fix
I'm honestly wondering why this hasn't been updated. It was very easy to do, as I just had to look at SpriteViewMatrix.cs in the source code and just make minor tweaks. I also added a comment with some code showing how to do things with ModifyTransformMatrix that the Volcano Tremor doesn't implement. (Namely, zooming the screen and flipping it)