Added a menu for when world gen crashes. (#825)
* Added a menu for when world gen crashes.

Added a menu for when world gen crashes that shows you the name of the faulty pass and the expection message. Also includes an Open Logs button and a button to return back to the main menu.

* World Gen crash now uses UIErrorMessage

* Open Logs button no longer pseudo-softlocks when clicked.

Open Log button used to pass a file path to a method expecting a directory path resulting in an uncaught exception and being unable to move the mouse untill resizing the window. Now it passes the directory instead of the file.

* Reverted json changes.

* Changed Open Logs button to use Process.Start