Two Lines to Fix Bestiary UI
ContentSamples probably needs to reconstruct the Bestiary Sort IDs after every Initialize,  because the Sort IDs are cleared.

It seems that the last time 'Initialize' is called is in ModContent's Load, so potentially, only that line is required.
However, since im unfamiliar with how mods load/unload, I've also patched it in Unload.