Potential Fix for water CTD (#1080)
It looks like the game expects a 13th water asset to be loaded, but it isn't being loaded.  This causes a NullReferenceException in some Draw code which causes a CTD.

After looking at the asset location, there is a 13th water source (in 1.4.0.5 at least), so this patch should add it.

Incase anyone would like to replicate the issue.
Seed: 1.4.1.1924359672 (Journey mode)
Go left and enter the underground tunnels, when you reach a fork, keep going left and up; you should see the beginnings of the underground desert.

Without the patch, it crashes, with the patch, there doesnt seem to be an issue.