Accessory Fix
This was a doozy to pinpoint.

Basically, Terraria had a bunch of checks on whether IDs for mods were valid sprites. This works great for vanilla, but any mods that added accessories didnt work. (Save for shield and wings, which I believe were patched previously)

This tries to patch as many of the checks as I could find with appropriate length values.

For consistency, I also readded the checks to wings and shields