Other slots also updated
I was reluctant to do this at first, but after looking at EquipTypes, and realizing body is probably clothing; I figured it would be a good idea to expand this to the other layers.

I may have missed one or flubbed something.
Especially with `flag3 = drawinfo.drawPlayer.body > 0;`
I think that just draws the default clothes if nothing is being worn, but if its only meant to be drawn by vanilla clothing, then I removed the check for that.