1.4 Potential Fix #1087 (#1090)
* Added Example Network Item

Trying to test #1087 on 1.4 branch
Couldn't reproduce, will need to ask for more info later

* Potential Fix

* Removing DebugNetworkingModWorld

It's a hot mess I made to test things, not really an example of how to use Networking for a ModWorld.

* Revert "Removing DebugNetworkingModWorld"

This reverts commit 5a20b714893db4c1afc734ee8cc3961eb809c959.

* Revert "Revert "Removing DebugNetworkingModWorld""

This reverts commit 333dee16adca4aba37f4717eaac8dea91be56b0c.

* Example of Solution 3 Mentioned in PR #1090

I dont like this because the logic in NetMessage seems very arbitrary. Also Clients[remoteClient] isnt synced, it just seems like an accident waiting to happen.

* Revert "Example of Solution 3 Mentioned in PR #1090"

This reverts commit 69e0da44b7f16810a5d924524614af0b94cd51e7.

* Avoid MessageBuffer Corruption Fix

My Previous fix would cause corruption when the client was not yet in state 4; while this would be fixed in state 10, it may have unintended side effects. This should hopefully remove any unintented side effects.