[1.4] Update Spanish localization (#1664)
* [1.4] Update Spanish localization

I have added the Steam Workshop related strings, as well as modified line 288 to make it fit on the in-game window.

* Update tModLoader.json

Changed a few more strings to make the translations more consistent. Made string 34 shorter (as it seems the text is cut if it is too long).

* Mistake

I accidentally deleted some words.