AwesomePrefix -> ExamplePrefix (#1441)
* AwesomePrefix -> ExamplePrefix

Added `ExamplePrefix` to ExampleMod based on the old `AwesomePrefix`

* Updated ExamplePrefix

It turns out the code is really against loading 2 similar prefixes in 1 class. Also one of the methods I was using was deprecated.

* Updated ExamplePrefix

Re-added autoloading multiple prefixes, with a nested loader class (which does get autoloaded) to load the prefixes themselves (Which don't get autoloaded)

* Fixed weird indentation

* Updated ExamplePrefix some more

Changed the loader from a `ModType` to an `ILoadable`. Also added small display name spacing.

* Adjust a few comments, and rename _displayName to _name

* _displayName -> _name

* Redone with inheritance, finalized.

Co-authored-by: Chicken-Bones <mehvids@gmail.com>
Co-authored-by: Mirsario <me@mirsar.io>