Fix bucket duplication bugs
This should fix #2077 and #1932

The issue was that there was no check for the stack of buckets being
empty after use. I don't know if this is fundamentally a symptom of some
deeper issue, but this is a small patch that solves the problem by
simply setting the item to air if the stack is empty at the end of
ItemCheck_UseBuckets.