Fix bucket duplication glitches
Second attempt at fixing this, done in a slightly better way. This makes
it so that itemAnimation is not reset if the stack size is 0, which is
important to prevent using nothing forever. This is vanilla behavior
broken by the item use change.

It also adds a check for stackSize <= 0 to useBucket. This is the way
vanilla does it for other items. It was left out for buckets probably
because it happened to work without it, which I'm pretty sure was a pure
coincidence