[1.4] Update Spanish localization (#2467)
* [1.4] Update Spanish localization

Updated the translations with the ones that appear at the English file (except for the placeholders). Tweaked some of the older ones.

* Update tModLoader.json

Shortened a translation that was too long for the button where it is in-game.

* Fix last commit

I added an unnecessary word, sorry. This is the last update.

* Resolve conflict

Added a new string from another commit and tweaked other translations to fit better in the UI and look better in Spanish.

* formatting fix

Co-authored-by: Igmc4 <31896129+Igmc@users.noreply.github.com>
Co-authored-by: JavidPack <javidpack@gmail.com>