Fixes 1
Fixed a spacing issue in IngameOptions.cs caused by skipping the drawing of the achievements option incorrectly.

Fixed a capitalization error: "Steam Workshop page" to "Steam Workshop Page".

Add localization keys in the lang files for the "Next Page", "Previous Page", etc. in the informational accessories pages.

Fixed the spread of the shadow in Utils.DrawBorderStringBig.