Resolves #3221
There was a bug with map files having the potential to create files
named as reserved keywords/files in Windows, which causes the file to
become incredibly difficult to remove.

This patch simply adds a check to the world and player names to check if
they have any of these reserved items and appends a '_' to them.