Remix Hallow worldgen is now in the "Hardmode Good" genpass
### What is the new feature?
Remix hallow worldgen now is properly included in the Hardmode Good genpass that tmod adds. This also makes the Hardmode Evil also run in remix seeds aswell

### Why should this be part of tModLoader?
This allows worldgen that relies on either one of the Hardmode Good and Hardmode Evil to be ran in remix seeds.

### Are there alternative designs?
Not that I can identify, the current patch for the remix seed was a bad fix when porting tmodloader to 1.4.4

### Sample usage for the new feature
Made a new mod, make worldgen that relies on the Hardmode Good pass to run. Results on the remix seed