Change SplitStack() favorited behavior to match OnStack()
OnStack() causes the favorited status to follow the item to the destination item, where SplitStack() removes favorited status from the destination item.  
Example:
Stack of 4 throwing knives in your inventory that is favorited.
Right click the stack once, now 1 in the mouse, 3 in your inventory.  The stack in your inventory is favorited and the on in the mouse is not.  Right click a 2nd time, and the inventory stack looses favorited and the mouse stack gains favorited.