MapIO liquid type changes
Reverted the MapIO to my previous changes back in 2023 using the fallback of not finding wall data with the same mod + content piece name to then get liquid data.

I'd prefer a better solution although BinaryReader doesn't give me too much to work with in terms of checking whether a map's data is modern or not

Currently this has the issue where if a mod has a wall and liquid of the same name, then it grabs the wall's map entry instead.